+++ src/gateways/Vimeo.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @link      https://dukt.net/videos/
- * @copyright Copyright (c) 2019, Dukt
+ * @copyright Copyright (c) 2020, Dukt
  * @license   https://github.com/dukt/videos/blob/v2/LICENSE.md
  */
 
@@ -10,6 +10,7 @@ namespace dukt\videos\gateways;
 use dukt\videos\base\Gateway;
 use dukt\videos\errors\CollectionParsingException;
 use dukt\videos\errors\VideoNotFoundException;
+use dukt\videos\helpers\VideosHelper;
 use dukt\videos\models\Collection;
 use dukt\videos\models\Section;
 use dukt\videos\models\Video;
@@ -476,7 +477,6 @@ class Vimeo extends Gateway
         $video->authorName = $data['user']['name'];
         $video->authorUrl = $data['user']['link'];
         $video->date = new DateTime($data['created_time']);
-        $video->durationSeconds = $data['duration'];
         $video->description = $data['description'];
         $video->gatewayHandle = 'vimeo';
         $video->gatewayName = 'Vimeo';
@@ -486,6 +486,10 @@ class Vimeo extends Gateway
         $video->url = 'https://vimeo.com/'.substr($data['uri'], 8);
         $video->width = $data['width'];
         $video->height = $data['height'];
+
+        // Video duration
+        $video->durationSeconds = $data['duration'];
+        $video->duration8601 = VideosHelper::getDuration8601($data['duration']);
 
         $this->parsePrivacy($video, $data);
         $this->parseThumbnails($video, $data);
@@ -527,23 +531,17 @@ class Vimeo extends Gateway
         }
 
         $largestSize = 0;
-        $thumbSize = 0;
 
         foreach ($this->getVideoDataPictures($data, 'thumbnail') as $picture) {
             // Retrieve highest quality thumbnail
             if ($picture['width'] > $largestSize) {
-                $video->thumbnailLargeSource = $picture['link'];
-                $largestSize = $picture['width'];
-            }
-
-            // Retrieve highest quality thumbnail with width < 400
-            if ($picture['width'] > $thumbSize && $thumbSize < 400) {
                 $video->thumbnailSource = $picture['link'];
-                $thumbSize = $picture['width'];
+                $largestSize = $picture['width'];
             }
         }
 
-        $video->thumbnailSource = $video->thumbnailSource ?? $video->thumbnailLargeSource;
+
+        $video->thumbnailLargeSource = $video->thumbnailSource;
 
         return null;
     }